물품 카드 애니메이션 및 카드 덱 구현 필요 : refactor : 카드 덱 위젯으로 추출 https://github.com/TEAM-ROMROM/RomRom-FE/issues/32
